requirements.txt
aiohttp==3.9.1 ; python_version >= '3.8'
aiohttp-retry==2.8.3 ; python_version >= '3.7'
aiosignal==1.3.1 ; python_version >= '3.7'
amqp==5.2.0 ; python_version >= '3.6'
annotated-types==0.6.0 ; python_version >= '3.8'
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.3 ; python_version < '3.11'
asyncssh==2.14.2 ; python_version >= '3.6'
atpublic==4.0 ; python_version >= '3.8'
attrs==23.2.0 ; python_version >= '3.7'
billiard==4.2.0 ; python_version >= '3.7'
celery==5.3.6 ; python_version >= '3.8'
certifi==2023.11.17 ; python_version >= '3.6'
cffi==1.16.0 ; python_version >= '3.8'
charset-normalizer==3.3.2 ; python_full_version >= '3.7.0'
click==8.1.7 ; python_version >= '3.7'
click-didyoumean==0.3.0 ; python_full_version >= '3.6.2' and python_full_version < '4.0.0'
click-plugins==1.1.1
click-repl==0.3.0 ; python_version >= '3.6'
colorama==0.4.6 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
configobj==5.0.8 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
contourpy==1.2.0 ; python_version >= '3.9'
cryptography==41.0.7 ; python_version >= '3.7'
cycler==0.12.1 ; python_version >= '3.8'
dictdiffer==0.9.0
diskcache==5.6.3 ; python_version >= '3'
distro==1.9.0 ; python_version >= '3.6'
dpath==2.1.6 ; python_version >= '3.7'
dulwich==0.21.7 ; python_version >= '3.7'
dvc==3.38.1
dvc-data==3.7.0 ; python_version >= '3.8'
dvc-http==2.32.0 ; python_version >= '3.8'
dvc-objects==3.0.3 ; python_version >= '3.8'
dvc-render==1.0.0 ; python_version >= '3.8'
dvc-studio-client==0.18.0 ; python_version >= '3.8'
dvc-task==0.3.0 ; python_version >= '3.8'
entrypoints==0.4 ; python_version >= '3.6'
filelock==3.13.1 ; python_version >= '3.8'
flatten-dict==0.4.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
flufl.lock==7.1.1 ; python_version >= '3.7'
fonttools==4.47.0 ; python_version >= '3.8'
frozenlist==1.4.1 ; python_version >= '3.8'
fsspec==2023.12.2 ; python_version >= '3.8'
funcy==2.0
gitdb==4.0.11 ; python_version >= '3.7'
gitpython==3.1.40 ; python_version >= '3.7'
grandalf==0.8
gto==1.6.1 ; python_version >= '3.8'
huggingface-hub==0.20.1 ; python_full_version >= '3.8.0'
hydra-core==1.3.2
idna==3.6 ; python_version >= '3.5'
iterative-telemetry==0.0.8 ; python_version >= '3.8'
jinja2==3.1.2 ; python_version >= '3.7'
kiwisolver==1.4.5 ; python_version >= '3.7'
kombu==5.3.4 ; python_version >= '3.8'
markdown-it-py==3.0.0 ; python_version >= '3.8'
markupsafe==2.1.3 ; python_version >= '3.7'
matplotlib==3.8.2
mdurl==0.1.2 ; python_version >= '3.7'
mpmath==1.3.0
multidict==6.0.4 ; python_version >= '3.7'
networkx==3.2.1 ; python_version >= '3.9'
numpy==1.26.3
omegaconf==2.3.0 ; python_version >= '3.6'
opencv-python==4.9.0.80
orjson==3.9.10 ; implementation_name == 'cpython'
packaging==23.2 ; python_version >= '3.7'
pathspec==0.12.1 ; python_version >= '3.8'
pillow==10.2.0
platformdirs==3.11.0 ; python_version >= '3.7'
prompt-toolkit==3.0.43 ; python_full_version >= '3.7.0'
psutil==5.9.7 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pycparser==2.21
pydantic==2.5.3 ; python_version >= '3.7'
pydantic-core==2.14.6 ; python_version >= '3.7'
pydot==2.0.0 ; python_version >= '3.7'
pygit2==1.13.3 ; python_version >= '3.8'
pygments==2.17.2 ; python_version >= '3.7'
pygtrie==2.5.0
pyparsing==3.1.1 ; python_full_version >= '3.6.8'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytorch-lightning==2.1.3
pyyaml==6.0.1
requests==2.31.0 ; python_version >= '3.7'
rich==13.7.0 ; python_full_version >= '3.7.0'
ruamel.yaml==0.18.5 ; python_version >= '3.7'
ruamel.yaml.clib==0.2.8 ; python_version < '3.13' and platform_python_implementation == 'CPython'
safetensors==0.4.1 ; python_version >= '3.7'
scmrepo==2.0.2 ; python_version >= '3.8'
semver==3.0.2 ; python_version >= '3.7'
setuptools==69.0.3 ; python_version >= '3.8'
shortuuid==1.0.11 ; python_version >= '3.5'
shtab==1.6.5 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smmap==5.0.1 ; python_version >= '3.7'
sqltrie==0.11.0 ; python_version >= '3.8'
sympy==1.12 ; python_version >= '3.8'
tabulate==0.9.0 ; python_version >= '3.7'
timm==0.9.12
tomlkit==0.12.3 ; python_version >= '3.7'
torch==2.1.2
torchvision==0.16.2 ; python_version >= '3.8'
tqdm==4.66.1 ; python_version >= '3.7'
typer==0.9.0 ; python_version >= '3.6'
typing-extensions==4.9.0 ; python_version >= '3.8'
tzdata==2023.4 ; python_version >= '2'
urllib3==2.1.0 ; python_version >= '3.8'
vine==5.1.0 ; python_version >= '3.6'
voluptuous==0.14.1 ; python_version >= '3.7'
wandb==0.16.1
wcwidth==0.2.12
yarl==1.9.4 ; python_version >= '3.7'
zc.lockfile==3.0.post1 ; python_version >= '3.7'
pytest
google-api-python-client
types-PyYAML